% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\name{Bank}
\alias{Bank}
\title{Connection to Bank website}
\description{
Create a class which will connect to a given bank when initialized and
download useful information which we want to store (balance, operations).
}
\section{Usage}{

\preformatted{
  bank <- Bank$new()
  bank$connect_driver(path_to_phantomjs = "")
  bank$connect_bank(bank, account_number, password)
  bank$get_balance()
  bank$get_operations(from = NULL, to = NULL)
  bank$disconnect_driver()
}
}

\section{Arguments}{

\describe{
  \item{bank}{Character, which bank to connect to}
  \item{account_number}{Numeric or character}
  \item{password}{Numeric or character}
  \item{path_to_phantomjs}{Character}
  \item{from}{Date}
  \item{to}{Date}
}
}

\section{Details}{

\code{$new()} creates a new instance of a bank object

\code{$connect_driver()} opens a web driver

\code{$connect_bank()} connects to a bank website

\code{$get_balance()} retrieves the balance for each account

\code{$get_operations()} retrieves operations from one date to another

\code{$disconnect_driver()} disconnect the web driver
}

