% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connexion_boursorama.R
\name{connect_boursorama}
\alias{connect_boursorama}
\title{connect_boursorama}
\usage{
connect_boursorama(account_number, password, path_to_phantomjs = "")
}
\arguments{
\item{account_number}{an integer corresponding to the cient account number}

\item{password}{an integer corresponding to the password which connect to the account}

\item{path_to_phantomjs}{a string}
}
\value{
The webpage get after the submit request
}
\description{
connect to the boursorama bank webpage
}

