% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_credit_agricole.R
\name{connect_credit_agricole}
\alias{connect_credit_agricole}
\title{connect_boursorama}
\usage{
connect_credit_agricole(driver, account_number, password, postal_code)
}
\arguments{
\item{driver}{a \code{\link[=RSelenium]{remoteDriver}} object}

\item{account_number}{a string corresponding to the cient account number}

\item{password}{a string corresponding to the password which connect to the account}

\item{postal_code}{an integer corresponding to the postal code of the bank (credit Agricole is a regionnal bank)}
}
\value{
The webpage get after the submit request
}
\description{
connect to the boursorama bank webpage
}
