% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_balance_boursorama.R
\name{get_balance_boursorama}
\alias{get_balance_boursorama}
\title{Get balance account}
\usage{
get_balance_boursorama(driver)
}
\arguments{
\item{driver}{a \code{\link[=RSelenium]{remoteDriver}} object}
}
\value{
a data frame
}
\description{
Fetch from website the balance of every account registered
}
