% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_operations_boursorama.R
\name{get_operations_boursorama}
\alias{get_operations_boursorama}
\title{Download operations from Boursorama}
\usage{
get_operations_boursorama(driver, from = as.Date("2000-01-01"),
  to = Sys.Date(), by = "1 year")
}
\arguments{
\item{driver}{a \code{\link[=RSelenium]{remoteDriver}} object}

\item{from, to}{a Date object}

\item{by}{increment of the sequence}
}
\value{
a data frame
}
\description{
Download operations from Boursorama bank on all account (except savings)
}
